requirements.txt
httpx[http2]>=0.27.0
telethon>=1.36.0
python-dotenv>=1.0.0
natsort>=8.4.0
Pillow>=10.0.0
cryptg>=0.4.0
